Time: 44 ms (50.47%), Space: 77.7 MB (23.22%) - LeetHub

+++ 1833-maximum-ice-cream-bars/1833-maximum-ice-cream-bars.java
@@ -2,10 +2,10 @@ class Solution {
 
     public int maxIceCream(int[] costs, int coins) {
         Arrays.sort(costs);
-        long cost = 0;
+        int cost = 0;
         int count = 0;
         for (int i = 0; i < costs.length; i++) {
-            if ((long) (costs[i]) + cost <= coins) {
+            if ( (costs[i]) + cost <= coins) {
                 cost += costs[i];
                 count++;
             } else {
